New programes added

+++ Cod Cheff/BINARYSUB.cpp
@@ -10,11 +10,11 @@ ll MOD=1000000007;
 void solve() {
     string s;
     cin >> s;
-    int n = s.length();
+    ll n = s.length();
     vi dp(n+1,0);
     dp[n]=1;
     dp[n-1] = 1;
-    for(int i=n-2;i>=0;i--) {
+    for(ll i=n-2;i>=0;i--) {
         if(s[i]!=s[i+1]) {
             dp[i]=dp[i+1]+dp[i+2]%998244353;
         } else {
@@ -26,7 +26,7 @@ void solve() {
  
  
  
-int main() {
+ll main() {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
     ll t;
